Moved sub Accordion interfaces to typef ile to avoid circular dependency
